优化
去掉controller层的try-catch，优化全局异常捕获处理器
优化请求参数的验证，减少代码量

+++ imba-radish/src/main/java/com/soft/base/mapper/SysMenuMapper.java
@@ -26,9 +26,9 @@ public interface SysMenuMapper extends BaseMapper<SysMenu> {
 
     List<Long> getByIds(@Param("ids") List<Long> ids);
 
-    void enableMenu(@Param("id") Integer id);
+    void enableMenu(@Param("id") Long id);
 
-    void disableMenu(@Param("id") Integer id);
+    void disableMenu(@Param("id") Long id);
 
     List<GetSelectMenuVo> getSelectMenu(@Param("type") String type);
